Slow hero leveling to a months-long chase; widen XP pipeline to long
Hero level stays the power/build-depth axis; just paced for the long haul.

- XpCurve steepened: XpBaseCost(600) * XpGrowth(1.19)^(level-1) -> ~95B XP total
  to level 100 (~140x the old 1.15 curve). Stays gentle through the early
  skill-unlock levels (~5-18), brutal in the back half.
- XP pipeline widened int -> long (HeroInstance.Xp, CombatState.PendingXp,
  GrantXp/GrantPartyXp amount), and the idle grant no longer clamps at 2.1B per
  claim (Idle.cs) -- which was silently losing deep-stage idle XP. Deep levels
  now reach into the billions safely.

349 GameCore tests (curve guard + overflow-safety + updated XP expectations).
Pacing (12h/day idle-equiv): ~stage 70 farm ≈ 4 months to 100; deeper is faster
(self-gating). Tunable via XpBaseCost/XpGrowth.

+++ gamecore/GameCore.Tests/CombatTests.cs
@@ -1201,7 +1201,7 @@ namespace IdleGame.GameCore.Tests
             Combat.RunToEnd(s, Cfg, new Rng(1));
 
             Assert.Equal(CombatStatus.Won, s.Status);
-            Assert.Equal(Cfg.Monsters["slime"].XpReward + Cfg.Monsters["goblin_king"].XpReward, s.PendingXp);
+            Assert.Equal((long)(Cfg.Monsters["slime"].XpReward + Cfg.Monsters["goblin_king"].XpReward), s.PendingXp);
         }
 
         [Fact]
@@ -1211,7 +1211,7 @@ namespace IdleGame.GameCore.Tests
                 Ent("A", Team.Party, hp: 1000, atk: 500, def: 0),
                 Ent("B", Team.Enemy, hp: 10, atk: 0, def: 0)); // RefKind "test"
             Combat.RunToEnd(s, Cfg, new Rng(1));
-            Assert.Equal(0, s.PendingXp);
+            Assert.Equal(0L, s.PendingXp);
         }
 
         [Fact]
@@ -1244,7 +1244,7 @@ namespace IdleGame.GameCore.Tests
 
             var h1 = after.Heroes.Find(h => h.Id == "h1")!;
             Assert.Equal(s.PendingXp, h1.Xp);                    // goblin_king XP (60) < XpCurve(1) -> xp advances
-            Assert.Equal(0, save.Heroes.Find(h => h.Id == "h1")!.Xp); // original untouched
+            Assert.Equal(0L, save.Heroes.Find(h => h.Id == "h1")!.Xp); // original untouched
         }
 
         // --- M4.3: hero downing + respawn ---
